feat: refactored component structure for sidebar
